+++ Message/controller/SettingViewController.m
@@ -120,7 +120,7 @@
 - (UITableViewCell *)tableView:(UITableView *)tableView cellForRowAtIndexPath:(NSIndexPath *)indexPath{
    
     UITableViewCell *cell = nil;
-    NSLog(@"%d,%d",indexPath.section,indexPath.row);
+    NSLog(@"%zd,%zd",indexPath.section,indexPath.row);
     if (indexPath.section != kClearAllConversationSection) {
         if(indexPath.section == kNetStatusSection && indexPath.row == kNetStatusRow){
             cell  = [tableView dequeueReusableCellWithIdentifier:@"statuscell"];
@@ -168,7 +168,7 @@
     
      [tableView deselectRowAtIndexPath:indexPath animated:NO];
     
-    int cellTag = (indexPath.section + 1) *100 + indexPath.row;
+    NSInteger cellTag = (indexPath.section + 1) *100 + indexPath.row;
     switch (cellTag) {
         case kAboutCellTag:
         {
